Move toolbar actions to a dedicated namespace
