+++ src/Components/editor/toolbar/SharePlay.js
@@ -3,10 +3,10 @@ import { TouchableOpacity, Share } from 'react-native';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import * as Linking from 'expo-linking';
 
-import theme from '../../styles/theme.style';
-import I18n from '../../utils/i18n';
-import { showError } from '../../utils/flashMessage';
-import { upload } from '../../utils/firebase';
+import theme from '../../../styles/theme.style';
+import I18n from '../../../utils/i18n';
+import { showError } from '../../../utils/flashMessage';
+import { upload } from '../../../utils/firebase';
 
 const SharePlay = ({ currentPlay }) => {
   const share = async () => {
